:recycle: [#297] Refactor custom field types
