+++ src/openforms/conf/base.py
@@ -190,6 +190,7 @@ INSTALLED_APPS = [
     "openforms.payments.contrib.ogone.apps.OgoneApp",
     "openforms.data_removal",
     "openforms.utils",
+    "openforms.custom_field_types",
 ]
 
 MIDDLEWARE = [
