+++ src/openforms/contrib/brp/apps.py
@@ -3,6 +3,3 @@ from django.apps import AppConfig
 
 class BRPConfig(AppConfig):
     name = "openforms.contrib.brp"
-
-    def ready(self):
-        from . import field_types  # noqa
