+++ src/openforms/custom_field_types/admin.py
@@ -1,0 +1,10 @@
+from django.contrib import admin
+
+from solo.admin import SingletonModelAdmin
+
+from .models import FamilyMembersTypeConfig
+
+
+@admin.register(FamilyMembersTypeConfig)
+class FamilyMembersTypeConfigAdmin(SingletonModelAdmin):
+    pass
