+++ src/openforms/fixtures/default_admin_index.json
@@ -217,6 +217,10 @@
             [
                 "zgw_consumers",
                 "service"
+            ],
+            [
+                "custom_field_types",
+                "familymemberstypeconfig"
             ]
         ]
     }
